pyproject.toml
[project]
name = "fluidsvelte"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.69.0",
    "grpcio-tools>=1.69.0",
]
